スコア登録処理をリファクタリング
disconnectするタイミングだけ、データの取得が終わった後に変わっている
